+++ CommonTools/include/MuonOutputTool.h
@@ -13,19 +13,19 @@
 
 #include "SusyAnalysisTools/include/SusyEnums.h"
 
-// ============================================================================
+// =============================================================================
 // Forward declare SusyDiLeptonCutFlowCycle
 class SusyDiLeptonCutFlowCycle;
 class SusyDiLeptonChargeFlipCycle;
-// ============================================================================
+// =============================================================================
 namespace CommonTools
 {
-// ============================================================================
+// =============================================================================
   class MuonOutputTool : public ToolBase
   {
-  // ----------------------------------------------------------------------------
+  // ---------------------------------------------------------------------------
   public:
-    // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+    // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
     MuonOutputTool(SCycleBase* parent, const char* name = "OutputTool");
     virtual ~MuonOutputTool();
 
@@ -40,7 +40,7 @@ namespace CommonTools
                    , VertexContainer&
                    );
 
-  // ----------------------------------------------------------------------------
+  // ---------------------------------------------------------------------------
   private:
 
     bool c_do_detailed_output;
@@ -49,47 +49,45 @@ namespace CommonTools
 
     //Variables to always output
 
-    std::vector<double>       m_mu_pt;
-    std::vector<double>       m_mu_eta;
-    std::vector<double>       m_mu_phi;
-    std::vector<double>       m_mu_cov_d0;
-    std::vector<double>       m_mu_d0;
-    std::vector<double>       m_mu_z0;
-    std::vector<double>       m_mu_d0_sig;
-    std::vector<double>       m_mu_z0_sin_theta;
-    std::vector<int>          m_mu_charge;
-    std::vector<double>       m_mu_etcone20;
-    std::vector<double>       m_mu_etcone30;
-    std::vector<double>       m_mu_etcone40;
-    std::vector<double>       m_mu_ptcone20;
-    std::vector<double>       m_mu_ptcone30;
-    std::vector<double>       m_mu_ptcone40;
-
+    std::vector<float> m_mu_pt;
+    std::vector<float> m_mu_eta;
+    std::vector<float> m_mu_phi;
+    std::vector<float> m_mu_cov_d0;
+    std::vector<float> m_mu_d0;
+    std::vector<float> m_mu_z0;
+    std::vector<float> m_mu_d0_sig;
+    std::vector<float> m_mu_z0_sin_theta;
+    std::vector<short> m_mu_charge;
+    std::vector<float> m_mu_etcone20;
+    std::vector<float> m_mu_etcone30;
+    std::vector<float> m_mu_etcone40;
+    std::vector<float> m_mu_ptcone20;
+    std::vector<float> m_mu_ptcone30;
+    std::vector<float> m_mu_ptcone40;
 
     std::vector<ull_t> m_mu_desc;
 
-
     // Detailed Variables to output only if configured to 
 
-    std::vector<int>          m_mu_b_layer_hits;
-    std::vector<int>          m_mu_num_pix_hits;
-    std::vector<int>          m_mu_num_sct_hits;
-    std::vector<int>          m_mu_num_trt_hits;
-    std::vector<int>          m_mu_expect_b_layer_hit;
-    std::vector<int>          m_mu_num_pix_holes;
-    std::vector<int>          m_mu_num_sct_holes;
-    std::vector<int>          m_mu_num_trt_outliers;
-    std::vector<int>          m_mu_num_trt_ht_outliers;
-    std::vector<int>          m_mu_num_mdt_layers;
+    std::vector<int> m_mu_b_layer_hits;
+    std::vector<int> m_mu_num_pix_hits;
+    std::vector<int> m_mu_num_sct_hits;
+    std::vector<int> m_mu_num_trt_hits;
+    std::vector<int> m_mu_expect_b_layer_hit;
+    std::vector<int> m_mu_num_pix_holes;
+    std::vector<int> m_mu_num_sct_holes;
+    std::vector<int> m_mu_num_trt_outliers;
+    std::vector<int> m_mu_num_trt_ht_outliers;
+    std::vector<int> m_mu_num_mdt_layers;
 
-    std::vector<double>       m_mu_raw_etcone20;
-    std::vector<double>       m_mu_raw_etcone30;
-    std::vector<double>       m_mu_raw_etcone40;
-    std::vector<double>       m_mu_raw_ptcone20;
-    std::vector<double>       m_mu_raw_ptcone30;
-    std::vector<double>       m_mu_raw_ptcone40;
+    std::vector<float> m_mu_raw_etcone20;
+    std::vector<float> m_mu_raw_etcone30;
+    std::vector<float> m_mu_raw_etcone40;
+    std::vector<float> m_mu_raw_ptcone20;
+    std::vector<float> m_mu_raw_ptcone30;
+    std::vector<float> m_mu_raw_ptcone40;
 
-    // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+    // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
     ClassDef(MuonOutputTool, 0);
   };
 }
